LPS-102367 Run "yarn install" if the project's node_modules directory does not exist
